+++ torchbit/debug/__init__.py
@@ -1,2 +1,3 @@
 from .judge import compare
 from .signallist import Signal, SignalGroup, generate_gtkwave_tcl
+from .bit_struct_comparator import compare_bit_struct